walletd contract, few more methods and basic error handling

+++ src/main/java/cash/ird/walletd/model/body/Status.java
@@ -5,12 +5,12 @@ import lombok.Data;
 @Data
 public class Status {
 
-    private int peerCount;
+    private long peerCount;
 
-    private int blockCount;
+    private long blockCount;
 
     private String lastBlockHash;
 
-    private int knownBlockCount;
+    private long knownBlockCount;
 
 }
